tried to make stable pointes better

+++ c++/include/vectorFuncs.hpp
@@ -23,7 +23,7 @@ std::ostream& operator<<(std::ostream& os, const std::vector<S>& vector){
 
 vector<vector<int>> vector1Dto2D(vector<int> input, vector<size_t> newShape);
 
-bool isClose(int* pointA, int* pointB, int thresh);
+bool isClose(int* pointA, int* pointB, double thresh);
 
 void avrageVectorValues(int dest[4], int** vector, int locA, int locB);
